finish build module, next step will test module

+++ Common/Header/millisecond.h
@@ -6,6 +6,6 @@
 
 void MilliSecond_Setup(void);
 void MilliSecond_Start(void);
-uint8_t MilliSecond_Stop(void);
+uint16_t MilliSecond_Stop(void);
 
 #endif
